added alert messages

+++ src/components/home/Home.jsx
@@ -4,6 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import { addDoc, collection, deleteDoc, doc, getDocs, updateDoc } from 'firebase/firestore';
 import { db } from '../../firebase-config';
 import Loader from '../loader/Loader';
+import Swal from 'sweetalert2';
 
 const Home = () => {
     const userName = localStorage.getItem('googleUserName');
@@ -51,10 +52,27 @@ const Home = () => {
     }, [refreshList, userEmail]);
 
     const handleSignOut = () => {
-        localStorage.removeItem('googleUserName');
-        localStorage.removeItem('googleUserEmail');
-        localStorage.removeItem('googleUserPhotoUrl');
-        navigate('/login');
+        Swal.fire({
+            title: 'Sign out?',
+            text: "All your data will remain persistent",
+            icon: 'error',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, sign out!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                localStorage.removeItem('googleUserName');
+                localStorage.removeItem('googleUserEmail');
+                localStorage.removeItem('googleUserPhotoUrl');
+                navigate('/login');
+                Swal.fire(
+                    'Signed out!',
+                    'You have been signed out successfully',
+                    'success'
+                )
+            }
+        });
     };
 
     const handleEdit = async (id) => {
@@ -65,9 +83,26 @@ const Home = () => {
     };
 
     const handleDelete = async (id) => {
-        const wordDoc = doc(db, userEmail, id);
-        await deleteDoc(wordDoc);
-        setRefreshList(true);
+        Swal.fire({
+            title: 'Delete message?',
+            text: "You won't be able to revert this!",
+            icon: 'error',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                const wordDoc = doc(db, userEmail, id);
+                await deleteDoc(wordDoc);
+                setRefreshList(true);
+                Swal.fire(
+                    'Deleted!',
+                    'Your message has been deleted successfully',
+                    'success'
+                )
+            }
+        });
     };
 
     const handleAddWord = async () => {
@@ -129,42 +164,42 @@ const Home = () => {
                                 <Fragment>
                                     {
                                         sortedData.length > 0 ?
-                                        <Fragment>
-                                            {
-                                                sortedData.map((word) => (
-                                                    <div className={classes.singleElement} key={word.id}>
-                                                        <span className={`${classes.activityName} ${editItemId === word.id ? classes.hideLayout : ''}`}>{word.message}</span>
-                                                        <div className={editItemId === word.id ? '' : classes.hideLayout}>
-                                                            <div className={classes.activityNameEdit}>
-                                                                <textarea defaultValue={word.message} className={classes.activityName} onChange={(e) => setEditedText(e.target.value)} />
-                                                                <div className={classes.editButtonsLayout}>
-                                                                    <button className={classes.editButton} onClick={() => handleEdit(word.id)}>Edit</button>
-                                                                    <button className={classes.editCancelButton} onClick={() => setEditItemId(null)}>Cancel</button>
+                                            <Fragment>
+                                                {
+                                                    sortedData.map((word) => (
+                                                        <div className={classes.singleElement} key={word.id}>
+                                                            <span className={`${classes.activityName} ${editItemId === word.id ? classes.hideLayout : ''}`}>{word.message}</span>
+                                                            <div className={editItemId === word.id ? '' : classes.hideLayout}>
+                                                                <div className={classes.activityNameEdit}>
+                                                                    <textarea defaultValue={word.message} className={classes.activityName} onChange={(e) => setEditedText(e.target.value)} />
+                                                                    <div className={classes.editButtonsLayout}>
+                                                                        <button className={classes.editButton} onClick={() => handleEdit(word.id)}>Edit</button>
+                                                                        <button className={classes.editCancelButton} onClick={() => setEditItemId(null)}>Cancel</button>
+                                                                    </div>
+                                                                </div>
+                                                            </div>
+                                                            <div className={classes.rightElements}>
+                                                                <span className={classes.activityDateTime}>{getFormattedDate(word.date)}</span>
+                                                                <div className={classes.modifyIcons}>
+                                                                    <div className={classes.editMessage} onClick={() => setEditItemId(word.id)}>
+                                                                        <svg style={{ cursor: 'pointer' }} xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-pencil-square" viewBox="0 0 16 16">
+                                                                            <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z" />
+                                                                            <path fillRule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z" />
+                                                                        </svg>
+                                                                    </div>
+                                                                    <div className={classes.deleteMessage} onClick={() => handleDelete(word.id)}>
+                                                                        <svg style={{ cursor: 'pointer' }} xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-x-circle-fill" viewBox="0 0 16 16">
+                                                                            <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zM5.354 4.646a.5.5 0 1 0-.708.708L7.293 8l-2.647 2.646a.5.5 0 0 0 .708.708L8 8.707l2.646 2.647a.5.5 0 0 0 .708-.708L8.707 8l2.647-2.646a.5.5 0 0 0-.708-.708L8 7.293 5.354 4.646z" />
+                                                                        </svg>
+                                                                    </div>
                                                                 </div>
                                                             </div>
                                                         </div>
-                                                        <div className={classes.rightElements}>
-                                                            <span className={classes.activityDateTime}>{getFormattedDate(word.date)}</span>
-                                                            <div className={classes.modifyIcons}>
-                                                                <div className={classes.editMessage} onClick={() => setEditItemId(word.id)}>
-                                                                    <svg style={{ cursor: 'pointer' }} xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-pencil-square" viewBox="0 0 16 16">
-                                                                        <path d="M15.502 1.94a.5.5 0 0 1 0 .706L14.459 3.69l-2-2L13.502.646a.5.5 0 0 1 .707 0l1.293 1.293zm-1.75 2.456-2-2L4.939 9.21a.5.5 0 0 0-.121.196l-.805 2.414a.25.25 0 0 0 .316.316l2.414-.805a.5.5 0 0 0 .196-.12l6.813-6.814z" />
-                                                                        <path fillRule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z" />
-                                                                    </svg>
-                                                                </div>
-                                                                <div className={classes.deleteMessage} onClick={() => handleDelete(word.id)}>
-                                                                    <svg style={{ cursor: 'pointer' }} xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-x-circle-fill" viewBox="0 0 16 16">
-                                                                        <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zM5.354 4.646a.5.5 0 1 0-.708.708L7.293 8l-2.647 2.646a.5.5 0 0 0 .708.708L8 8.707l2.646 2.647a.5.5 0 0 0 .708-.708L8.707 8l2.647-2.646a.5.5 0 0 0-.708-.708L8 7.293 5.354 4.646z" />
-                                                                    </svg>
-                                                                </div>
-                                                            </div>
-                                                        </div>
-                                                    </div>
-                                                ))
-                                            }
-                                        </Fragment>
-                                        :
-                                        <div className={classes.noDataFound}>No data found</div>
+                                                    ))
+                                                }
+                                            </Fragment>
+                                            :
+                                            <div className={classes.noDataFound}>No data found</div>
                                     }
                                 </Fragment>
                         }
